feat: Remove deprecated components and refactor URL input handling
