--- src/components/__init__.py
@@ -1,1 +1,0 @@
-# Components module for st-bubble-chat
