--- src/components/query_page/__init__.py
@@ -1,8 +1,0 @@
-from .chat_ui import render_ai_message, render_chat_messages
-from .query_page import render_query_page
-
-__all__ = [
-    "render_query_page",
-    "render_chat_messages",
-    "render_ai_message",
-]
