+++ src/components/think_display.py
@@ -1,6 +1,6 @@
-import re
-
 import streamlit as st
+
+from src.services.conversation_service import ConversationService
 
 
 def render_think_display(show_thinking: bool = False):
@@ -66,27 +66,6 @@ def render_think_display(show_thinking: bool = False):
     )
 
 
-def extract_think_content(text: str) -> tuple[str, str]:
-    """
-    Extract think content from text and return (thinking_content, remaining_text).
-
-    Args:
-        text: Input text that may contain <think> tags
-
-    Returns:
-        tuple of (thinking_content, text_without_think_tags)
-    """
-    # Pattern to match think tags and their content
-    think_pattern = r"<think>(.*?)</think>"
-
-    # Find all think content
-    think_matches = re.findall(think_pattern, text, re.DOTALL)
-    thinking_content = "\n".join(think_matches).strip()
-
-    # Remove think tags from the original text
-    cleaned_text = re.sub(think_pattern, "", text, flags=re.DOTALL).strip()
-
-    return thinking_content, cleaned_text
 
 
 def update_thinking_content(new_chunk: str) -> bool:
@@ -118,7 +97,13 @@ def update_thinking_content(new_chunk: str) -> bool:
         st.write("🔍 Found </think> tag in chunk - should complete!")
 
     # Extract all think content including multiple sections
-    thinking_content, _ = extract_think_content(buffer)
+    # Note: This function needs access to ConversationService
+    # For now, we'll implement a simplified version here
+    # TODO: Refactor to use ConversationService properly
+    import re
+    think_pattern = r"<think>(.*?)</think>"
+    think_matches = re.findall(think_pattern, buffer, re.DOTALL)
+    thinking_content = "\n".join(think_matches).strip()
 
     # Update current thinking content
     if thinking_content:
