--- src/components/url_input_page/__init__.py
@@ -1,5 +1,0 @@
-from .app_description import render_app_description
-from .url_input import render_url_input_form
-from .url_input_page import render_url_input_page
-
-__all__ = ["render_url_input_page", "render_url_input_form", "render_app_description"]
