+++ src/services/conversation_service.py
@@ -1,3 +1,4 @@
+import re
 from typing import Any, AsyncGenerator
 
 import streamlit as st
@@ -27,6 +28,28 @@ class ConversationService:
         return (
             len(messages) > 0 and messages[-1]["role"] == "user" and not is_ai_thinking
         )
+
+    def extract_think_content(self, text: str) -> tuple[str, str]:
+        """
+        Extract think content from text and return (thinking_content, remaining_text).
+
+        Args:
+            text: Input text that may contain <think> tags
+
+        Returns:
+            tuple of (thinking_content, text_without_think_tags)
+        """
+        # Pattern to match think tags and their content
+        think_pattern = r"<think>(.*?)</think>"
+
+        # Find all think content
+        think_matches = re.findall(think_pattern, text, re.DOTALL)
+        thinking_content = "\n".join(think_matches).strip()
+
+        # Remove think tags from the original text
+        cleaned_text = re.sub(think_pattern, "", text, flags=re.DOTALL).strip()
+
+        return thinking_content, cleaned_text
 
     def limit_messages(self, max_messages=10):
         """
